Notes
This is a piece of code that allows the user in the terminal to be able to create their own number list. Though this method takes as many milliseconds as the numbers you insert (Linear Time), it is a pretty good attempt on my behalf as I have just recently started learning Python.